# contracts/generated/openapi/pollen.v1.yaml
openapi: 3.1.0
info:
  title: Pollen DEK Control-Plane API
  version: 1.0.0
  contact:
    name: Pollen Platform Team
    url: https://your-cloud-domain.com
    email: platform@your-cloud-domain.com
  description: Contract between Pollen Cloud / Local Control Plane and the DEK edge enforcement plane.
...
paths:
...
  /v1/tenants/{tenant_id}/observations/costs:
    get:
      operationId: ObservationApi_getCosts
      description: Get cost summary
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PollenHeaders.contractVersion'
        - $ref: '#/components/parameters/PollenHeaders.deviceId'
        - $ref: '#/components/parameters/PollenHeaders.tenantId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostSummaryResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollenError'
      tags:
        - Observation
...
components:
  parameters:
...
    PollenHeaders.contractVersion:
      name: X-Pollen-Contract-Version
      in: header
      required: true
      schema:
        type: string
    PollenHeaders.deviceId:
      name: X-Pollen-Device-Id
      in: header
      required: false
      schema:
        type: string
    PollenHeaders.tenantId:
      name: X-Pollen-Tenant-Id
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    AgentUsageSummary:
      type: object
      required:
        - cost
        - tokens
      properties:
        cost:
          type: number
          format: float
        tokens:
          type: integer
          format: int64
...
    CostSummaryResponse:
      type: object
      required:
        - schema_version
        - tenant_id
        - period
        - total_estimated_cost_usd
        - total_tokens
        - provider_breakdown
        - agent_breakdown
        - agent_token_breakdown
        - agent_usage_breakdown
      properties:
        schema_version:
          type: string
          enum:
            - cost-summary.v1
        tenant_id:
          type: string
        period:
          type: string
        total_estimated_cost_usd:
          type: number
          format: float
        total_tokens:
          type: integer
          format: int64
        provider_breakdown:
          type: object
          unevaluatedProperties:
            type: number
            format: float
        agent_breakdown:
          type: object
          unevaluatedProperties:
            type: number
            format: float
        agent_token_breakdown:
          type: object
          unevaluatedProperties:
            type: integer
            format: int64
        agent_usage_breakdown:
          type: object
          unevaluatedProperties:
            $ref: '#/components/schemas/AgentUsageSummary'
...
    PollenError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        trace_id:
          type: string
...
servers:
  - url: http://127.0.0.1:17891
    description: Local Control Plane / Mock Cloud
    variables: {}
  - url: https://api.pollen.cloud
    description: Pollen Cloud
    variables: {}